AR working with Relay - Working for first player (host) moving balls, changing colors.
Working for first player (host) moving balls, changing colors.

*Run on Android.
Start game, first player Click Host, in main screen tap on a detected plane to place the table. Tap on the table or cup to move a ball.

Second player, enter the client code and click the client button.   in main screen tap on a detected plane to place the table. Tap on the table or cup to move a ball.

Ball movement not currently working for second player - need to verify the ownership settings on the gameobject's to allow movement.